pass the get API place to the backEnd
