+++ client/package.json
@@ -9,6 +9,7 @@
     "axios": "^0.15.3",
     "bootstrap": "^4.0.0-alpha.6",
     "classnames": "^2.2.5",
+    "jsonp": "^0.2.1",
     "jwt-decode": "^2.1.0",
     "normalizr": "^3.1.0",
     "ramda": "^0.23.0",
