+++ client/src/actions/auth.js
@@ -2,9 +2,10 @@ import R from 'ramda';
 import jwtDecode from 'jwt-decode';
 import api from '../api';
 import * as fromAuthHandler from '../utils/authTokenHandler';
-import {  normalizeUser} from '../normalizers';
+import {normalizeUser} from '../normalizers';
 import React from 'react';
 import http from 'axios';
+import jsonp from 'jsonp';
 
 export const signup = user => api.post('/users', user).then(null,
   ({
@@ -27,24 +28,18 @@ export const authenticateUser = dispatch => R.compose(
 
 export const login = user => dispatch => {
   return api.post('/authenticate', user).then(
-    ({
-      data
-    }) => {
+    ({data}) => {
       fromAuthHandler.storeAuthToken(data.token);
       authenticateUser(dispatch)(data.token);
     },
-    ({
-      response
-    }) => Promise.reject(response.data.errors),
+    ({response}) => Promise.reject(response.data.errors),
   );
 }
 
 export const logout = (id) => dispatch => {
   fromAuthHandler.removeAuthToken();
   fromAuthHandler.deleteAuthTokenRequest();
-  dispatch({
-    type: 'LOGOUT_USER'
-  });
+  dispatch({type: 'LOGOUT_USER'});
   dispatch({
     type: 'REMOVE_USER',
     userId: id
@@ -52,24 +47,12 @@ export const logout = (id) => dispatch => {
 }
 
 export const countryList = (name) => {
-  const key = 'AIzaSyBcASq82k5do_ZviitsV64QybYzsa-9O-E';
-   const url_city = `https://maps.googleapis.com/maps/api/place/autocomplete/json?input=${name}&types=geocode&language=en&key=${key}`;
-    http.get(url_city)
-        .then((response)=>{
-          if(response.status === 'OK'){
-              const data = response.predictions[0];
-              var Places = { city:"", country:""};
-              var descriptionPlace = [];
-              for(let item in data){
-                console.log(item.description);
-                descriptionPlace.push(item.description);
-              }
-              return  descriptionPlace;
-          }
-          else{ return "No match";}
+  //console.log(name);
 
-        })
-        .catch(()=>{return "Error come up with the API";})
+   const url_city = `https://maps.googleapis.com/maps/api/place/autocomplete/json?input=${name}`;
+   var data = api.get(`/users/getCountryList?url_city=${url_city}`);
+
+
   return {world:[{
                place:{
                  id:1,
